Görüntü_İncele/Görüntüleri_Detayli_İncele/requirements.txt
numpy
pandas
matplotlib
seaborn
pillow
scikit-learn
tqdm
